requirements.txt
flask==3.1.1
langchain==0.3.26
langchain-google-genai==2.0.8
langchain-pinecone==0.2.8
langchain-community==0.3.26
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
google-generativeai>=0.8.0
markdown==3.10
gunicorn==21.2.0